# api/openapi/api.oapi3.yaml
openapi: "3.0.0"
info:
 title: Simple service API
 description: REST API
 version: "1.0.0"
servers:
 - url: /
paths:
...
  /search/{q}:
    get:
      summary: Search item
      description: Search item
      operationId: findItems
      parameters:
        - name: q
          in: path
          description: item name from begin
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: array
                  items:
                    type: object
                    properties: {}
        400:
          description: bad request
        500:
          description: internal server error    